requirements.txt
aiogram==3.7.0
qdrant-client==1.7.0
sentence-transformers==2.5.1
numpy==1.26.4
python-dotenv==1.0.1
torch>=2.0.0
transformers>=4.30.0
pytest==8.0.0
pytest-asyncio==0.23.0 